renamed select-namespace to be namespace
