+++ src/commands/pipelines.ts
@@ -3,7 +3,7 @@ import { getCredentials } from '../services/codefresh/client.ts';
 import { getAccountRuntimes, getAccountRuntimeSpec } from '../services/codefresh/runtimes.ts';
 import { getResources } from '../services/kubernetes/resources.ts';
 import { collectData } from '../services/kubernetes/collector.ts';
-import { selectNamespace } from '../utils/select-namespace.ts';
+import { selectNamespace } from '../utils/namespace.ts';
 import { createDirPath, preparePackage, writeYaml } from '../utils/files.ts';
 
 export const pipelinesCommand = new Command()
